Fix array test
If after a literal there is a character like `]` then it should still parse as it is not part of the literal.
But if it is a normal character then it is an error and so it can be thrown.

Before it just checked if there is whitespace after that.